refactor: move all of paradigm.py -> paradigm_filler.py
refactor: move row/cell classes to paradigm_filler.py

remove: [root] paradigm.py

refactor: use Python 3.9 types instead of imported types

style: reorganize members
