+++ CreeDictionary/utils/paradigm_filler.py
@@ -1,29 +1,21 @@
+from __future__ import annotations
+
 """
-fill a paradigm table according to a lemma
+Fill a paradigm table with the inflections of a lemma.
 """
+
 import logging
 from copy import deepcopy
 from pathlib import Path
-from typing import Dict, List, Sequence, Set, Tuple, cast
+from string import Template
+from typing import Iterable, Literal, Optional, Sequence, Union, cast
 
+from attr import attrib, attrs
 from hfst_optimized_lookup import TransducerFile
-from paradigm import (
-    Cell,
-    EmptyRow,
-    InflectionCell,
-    Layout,
-    StaticCell,
-    TitleRow,
-    rows_to_layout,
-)
 from utils import ParadigmSize, WordClass, shared_res_dir
 from utils.types import ConcatAnalysis
 
 logger = logging.getLogger()
-
-RawTable = List[List[str]]
-LayoutTable = Dict[Tuple[WordClass, ParadigmSize], RawTable]
-LayoutID = Tuple[WordClass, ParadigmSize]
 
 PARADIGM_NAME_TO_WC = {
     "noun-na": WordClass.NA,
@@ -37,35 +29,17 @@ PARADIGM_NAME_TO_WC = {
 }
 
 
-def import_frequency() -> Dict[ConcatAnalysis, int]:
-    # TODO: store this in the database, rather than as a source file
-    # TODO: make a management command that updates wordform frequencies
-    FILENAME = "attested-wordforms.txt"
+##################################### Simple types #####################################
 
-    res: dict[ConcatAnalysis, int] = {}
-    lines = (shared_res_dir / FILENAME).read_text(encoding="UTF-8").splitlines()
-    for line in lines:
-        line = line.strip()
-        if not line:
-            # Skip empty lines
-            continue
+RawTable = list[list[str]]
+Layouts = dict[tuple[WordClass, ParadigmSize], RawTable]
+LayoutID = tuple[WordClass, ParadigmSize]
 
-        try:
-            freq, _, *analyses = line.split()
-        except ValueError:  # not enough value to unpack, which means the line has less than 3 values
-            logger.warn(f'line "{line}" is broken in {FILENAME}')
-        else:
-            for analysis in analyses:
-                res[ConcatAnalysis(analysis)] = int(freq)
 
-    return res
+##################################### Main classes #####################################
 
 
 class ParadigmFiller:
-    _layout_tables: Dict[LayoutID, Layout]
-    _generator: TransducerFile
-    _frequency = import_frequency()
-
     def __init__(self, layout_dir: Path, generator_hfstol_path: Path = None):
         """
         Combine .layout, .layout.csv, .paradigm files to paradigm tables of different sizes and store them in memory
@@ -74,6 +48,7 @@ class ParadigmFiller:
         :param layout_dir: the directory for .layout and .layout.cvs files
         """
         self._layout_tables = self._import_layouts(layout_dir)
+        self._frequency = import_frequency()
 
         if generator_hfstol_path is None:
             from shared import expensive
@@ -91,7 +66,7 @@ class ParadigmFiller:
 
     def fill_paradigm(
         self, lemma: str, category: WordClass, paradigm_size: ParadigmSize
-    ) -> List[Layout]:
+    ) -> list[Layout]:
         """
         Returns a paradigm table filled with word forms
 
@@ -105,12 +80,12 @@ class ParadigmFiller:
         # so set up some data structures that will allow us to:
         #  - store all unique things to lookup
         #  - remember which strings need to be replaced after lookups
-        lookup_strings: List[ConcatAnalysis] = []
-        string_locations: List[Tuple[List[Cell], int]] = []
+        lookup_strings: list[ConcatAnalysis] = []
+        string_locations: list[tuple[list[Cell], int]] = []
 
         layout_table = deepcopy(self._layout_tables[(category, paradigm_size)])
 
-        tables: List[Layout] = [[]]
+        tables: list[Layout] = [[]]
 
         for row in layout_table:
             if row is EmptyRow:
@@ -153,7 +128,7 @@ class ParadigmFiller:
 
     def inflect_all_with_analyses(
         self, lemma: str, wordclass: WordClass
-    ) -> Dict[ConcatAnalysis, Sequence[str]]:
+    ) -> dict[ConcatAnalysis, Sequence[str]]:
         """
         Produces all known forms of a given word. Returns a mapping of analyses to their
         forms. Some analyses may have multiple forms. Some analyses may not generate a
@@ -161,17 +136,17 @@ class ParadigmFiller:
         """
         analyses = self.expand_analyses(lemma, wordclass)
         return cast(
-            Dict[ConcatAnalysis, Sequence[str]], self._generator.bulk_lookup(analyses)
+            dict[ConcatAnalysis, Sequence[str]], self._generator.bulk_lookup(analyses)
         )
 
-    def inflect_all(self, lemma: str, wordclass: WordClass) -> Set[str]:
+    def inflect_all(self, lemma: str, wordclass: WordClass) -> set[str]:
         """
         Return a set of all inflections for a particular lemma.
         """
         all_inflections = self.inflect_all_with_analyses(lemma, wordclass).values()
         return set(form for word in all_inflections for form in word)
 
-    def expand_analyses(self, lemma: str, wordclass: WordClass) -> Set[ConcatAnalysis]:
+    def expand_analyses(self, lemma: str, wordclass: WordClass) -> set[ConcatAnalysis]:
         """
         Given a lemma and its word class, return a set of all analyses that we could
         generate, but do not actually generate anything!
@@ -183,7 +158,7 @@ class ParadigmFiller:
         layout_table = self._layout_tables[(wordclass, ParadigmSize.LINGUISTIC)]
 
         # Find all of the analyses strings in the table:
-        analyses: Set[ConcatAnalysis] = set()
+        analyses: set[ConcatAnalysis] = set()
         for row in layout_table:
             if row is EmptyRow or isinstance(row, TitleRow):
                 continue
@@ -203,7 +178,7 @@ class ParadigmFiller:
         return analyses
 
     @staticmethod
-    def _import_layouts(layout_dir) -> Dict[LayoutID, Layout]:
+    def _import_layouts(layout_dir) -> dict[LayoutID, Layout]:
         """
         Imports .layout files into memory.
 
@@ -222,8 +197,180 @@ class ParadigmFiller:
         return layout_tables
 
 
-def import_layouts(layout_file_dir: Path) -> LayoutTable:
-    layout_tables: LayoutTable = {}
+################################## Table/Pane classes ##################################
+
+
+class EmptyRowType:
+    """
+    A completely empty row! Singleton.
+    """
+
+    # Do a bunch of stuff to make this an empty row.
+    _instance: "EmptyRowType"
+
+    def __new__(cls) -> "EmptyRowType":
+        if not hasattr(cls, "_instance"):
+            cls._instance = super().__new__(cls)
+        return cls._instance
+
+    def __repr__(self) -> str:
+        return "EmptyRowType"
+
+    def __deepcopy__(self, _memo) -> "EmptyRowType":
+        return self
+
+    def __copy__(self) -> "EmptyRowType":
+        return self
+
+    def __reduce__(self):
+        return (EmptyRowType, ())
+
+
+EmptyRow = EmptyRowType()
+
+
+@attrs
+class TitleRow:
+    """
+    A row containing only a title -- no inflections of table headers.
+    """
+
+    title = attrib(type=str)
+    span = attrib(type=int)
+
+    # Makes it so that the Django template can easily determine that this is a
+    # title:
+    is_title = True
+
+
+class StaticCell:
+    """
+    A cell that undergoes no change in the when rendering a layout to a
+    paradigm.
+    """
+
+    is_heading: bool = False
+    is_label = False
+
+    text: str
+
+    def __str__(self) -> str:
+        return self.text
+
+
+@attrs(frozen=True)
+class Label(StaticCell):
+    """
+    A title in the rendered paradigm.
+    """
+
+    is_label = True
+    text = attrib(type=str)
+
+
+@attrs(frozen=True)
+class Heading(StaticCell):
+    """
+    A section header in the rendered paradigm.
+    """
+
+    is_heading = True
+    text = attrib(type=str)
+
+
+# frozen=False is a reminder that the inflection is default as None and generated later
+# inflection related info like inflection frequency is also generated later
+@attrs(frozen=False, auto_attribs=True, eq=False, repr=False)
+class InflectionCell:
+    # the analysis of the inflection (with the lemma to be filled out)
+    # It looks like for example "${lemma}+TAG+TAG+TAG", "TAG+${lemma}+TAG+TAG"
+    analysis: Optional[Template] = None
+
+    # the inflection to be generated
+    inflection: Optional[str] = None
+
+    # the frequency of the inflection in the corpus
+    frequency: Optional[int] = None
+
+    @property
+    def has_analysis(self):
+        return self.analysis is not None
+
+    def __eq__(self, other) -> bool:
+        # N.B., string.Template needs to be checked directly :/
+        return isinstance(other, InflectionCell) and (
+            (
+                self.analysis is not None
+                and other.analysis is not None
+                and self.analysis.template == other.analysis.template
+                and self.inflection == other.inflection
+                and self.frequency == other.frequency
+            )
+            or (self.analysis is None and other.analysis is None)
+        )
+
+    def __repr__(self) -> str:
+        if self.inflection or self.frequency and self.analysis is None:
+            return super().__repr__()
+
+        assert self.analysis is not None  # mypy is dumb...
+        return f"{type(self).__name__}(analysis=Template({self.analysis.template!r})"
+
+    def create_concat_analysis(self, lemma: str) -> ConcatAnalysis:
+        """
+        Fills in the analysis. Useful if you want to inflect this cell.
+
+        >>> cell = InflectionCell.from_raw_nds_cell("{{ lemma }}+V+II+Ind+3Sg")
+        >>> cell.create_concat_analysis("mispon")
+        'mispon+V+II+Ind+3Sg'
+        """
+        assert self.analysis is not None
+
+        return ConcatAnalysis(self.analysis.substitute(lemma=lemma))
+
+    @classmethod
+    def from_raw_nds_cell(cls, raw_cell: str) -> "InflectionCell":
+        """
+        Generates an InflectionCell from a NDS-style (legacy) template format.
+        """
+        return cls(Template(raw_cell.replace("{{ lemma }}", "${lemma}")))
+
+
+Cell = Union[InflectionCell, StaticCell, Literal[""]]
+Row = Union[list[Cell], EmptyRowType, TitleRow]
+# TODO: Make a class for a list of rows (a Pane)
+Layout = list[Row]
+
+
+################################## Internal functions ##################################
+
+
+def import_frequency() -> dict[ConcatAnalysis, int]:
+    # TODO: store this in the database, rather than as a source file
+    # TODO: make a management command that updates wordform frequencies
+    FILENAME = "attested-wordforms.txt"
+
+    res: dict[ConcatAnalysis, int] = {}
+    lines = (shared_res_dir / FILENAME).read_text(encoding="UTF-8").splitlines()
+    for line in lines:
+        line = line.strip()
+        if not line:
+            # Skip empty lines
+            continue
+
+        try:
+            freq, _, *analyses = line.split()
+        except ValueError:  # not enough value to unpack, which means the line has less than 3 values
+            logger.warn(f'line "{line}" is broken in {FILENAME}')
+        else:
+            for analysis in analyses:
+                res[ConcatAnalysis(analysis)] = int(freq)
+
+    return res
+
+
+def import_layouts(layout_file_dir: Path) -> Layouts:
+    layout_tables: Layouts = {}
 
     files = list(layout_file_dir.glob("*.layout.tsv"))
 
@@ -277,3 +424,90 @@ def parse_layout(layout_file: Path) -> RawTable:
         table.append(row)
 
     return table
+
+
+def rows_to_layout(rows: Iterable[list[str]]) -> Layout:
+    """
+    Takes rows (e.g., from a TSV file), and creates a well-formatted layout
+    file.
+    """
+    layout: Layout = []
+
+    max_row_length = len(max(rows, key=len))
+
+    for raw_row in rows:
+        row = determine_cells(raw_row)
+        row.extend([""] * (max_row_length - len(row)))
+
+        has_content = False
+        n_labels = 0
+        last_label: str
+
+        for cell in row:
+            if isinstance(cell, Label):
+                n_labels += 1
+                last_label = cell.text
+            elif cell != "":
+                has_content = True
+
+        if not has_content and n_labels == 0:
+            layout.append(EmptyRow)
+        elif not has_content and n_labels == 1:
+            layout.append(TitleRow(last_label, span=len(row)))
+        else:
+            layout.append(row)
+
+    return layout
+
+
+def does_raw_row_has_row_header(raw_row: list[str]) -> bool:
+    """
+    does the row host inflection or not?
+    """
+
+    # we check if the row has a "left aligned heading"
+
+    for i, raw_cell in enumerate(raw_row):
+
+        if (
+            raw_cell.startswith('"')
+            and raw_cell.endswith('"')
+            or raw_cell.startswith(":")
+        ):
+            return i == 0
+
+    return False
+
+
+def determine_cells(raw_row: list[str]) -> list[Cell]:
+    has_row_header = does_raw_row_has_row_header(raw_row=raw_row)
+    cells: list[Cell] = []
+
+    for raw_cell in raw_row:
+        cell: Cell
+
+        if raw_cell.startswith('"') and raw_cell.endswith('"'):
+            # This will be something like:
+            #   "Something is happening now"
+            #   "Speech act participants"
+            assert len(raw_cell) > 2
+            cell = Label(raw_cell[1:-1])
+        elif raw_cell.startswith(":"):
+            # This will be something like:
+            #   : "ê-/kâ- word"
+            #   : "ni-/ki- word"
+            _colon, content, _empty = raw_cell.split('"')
+            cell = Heading(content)
+        else:
+            raw_cell = raw_cell.strip()
+            if raw_cell == "":
+                if has_row_header:
+                    cell = InflectionCell()
+                else:
+                    cell = ""
+            else:
+                # "{{ lemma }}" is a proprietary format
+                cell = InflectionCell.from_raw_nds_cell(raw_cell)
+
+        cells.append(cell)
+    return cells
